@@ -1,7 +1,7 @@
 package com.hzy.cnn.CustomView.Test;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageView;
 
