@@ -1,8 +1,8 @@
 package com.hzy.cnn.CustomView.Test;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.hzy.cnn.CustomView.Ui.view.WaveView;
 import com.hzy.cnn.CustomView.R;
